pr/#80 완벽 반영
