+++ src/hooks/useGetBoardPosts.ts
@@ -1,24 +1,5 @@
 import { getBoardPosts } from '@/apis/getBoardPosts';
-import { useQuery, UseQueryResult } from '@tanstack/react-query';
-
-interface Post {
-  postId: number;
-  title: string;
-  content: string;
-  date: string;
-  category: string;
-  thumbNail: string;
-  status: string;
-}
-
-interface PageInfo {
-  totalPages: number;
-}
-
-interface GetBoardPostsResponse {
-  postListResDto: Post[];
-  pageInfo: PageInfo;
-}
+import { useQuery } from '@tanstack/react-query';
 
 interface UseGetBoardPostsProps {
   boardCode: string;
@@ -27,15 +8,10 @@ interface UseGetBoardPostsProps {
   category?: string | null;
 }
 
-export function useGetBoardPosts({
-  boardCode,
-  page = 0,
-  take = 9,
-  category = null,
-}: UseGetBoardPostsProps): UseQueryResult<GetBoardPostsResponse> {
+export function useGetBoardPosts({ boardCode, page = 0, take = 9, category = null }: UseGetBoardPostsProps) {
   const queryKey = ['get-board-boardCode-posts', boardCode, page, take, category];
 
-  return useQuery<GetBoardPostsResponse>({
+  return useQuery({
     queryKey,
     queryFn: () =>
       getBoardPosts({
