+++ src/hooks/usePostBoardFiles.ts
@@ -1,4 +1,5 @@
-import { postBoardFiles, postBoardFilesProps, PostBoardFilesResponse } from '@/apis/postBoardFiles';
+import { postBoardFiles, postBoardFilesProps } from '@/apis/postBoardFiles';
+import { PostBoardFilesResponse } from '@/types/apis/post';
 import { useMutation, UseMutationResult } from '@tanstack/react-query';
 import { AxiosError, AxiosResponse } from 'axios';
