--- src/types/apis/index.tsx
@@ -1,18 +1,0 @@
-export interface PostDetailResDto {
-  title: string;
-  createdAt: string;
-  content: string;
-  imageList: string[];
-}
-
-export interface GetBoardDetailResponse {
-  data: {
-    postDetailResDto: PostDetailResDto;
-  };
-}
-
-export interface getBoardDetailProps {
-  boardCode: string;
-  postId: number;
-  userId?: number;
-}
