+++ src/types/apis/post/index.tsx
@@ -1,0 +1,21 @@
+export interface FileData {
+  id: number;
+  url: string;
+}
+
+export interface PostBoardFilesResponse {
+  code: string;
+  message: string;
+  data: FileData[];
+  isSuccess: boolean;
+}
+
+export interface PostBoardPostsResponse {
+  code: string;
+  message: string;
+  data: {
+    post_id: number;
+    boardCode: string;
+  };
+  isSuccess: boolean;
+}
